implemented validation loss, scheduler
