+++ src/train_v1/config/model/torch_v1.yaml
@@ -5,10 +5,9 @@ model:
     hidden_size: 256
   train_param:
     # gpu_num: 8
-    batch_size: 16  # * gpu_num
+    batch_size: 64  # * gpu_num
     epochs: 200
     early_stopping_rounds: 30
-    # n_models: 1  # 5
 
 optimizer:
   name: 'Adam'
@@ -16,17 +15,28 @@ optimizer:
     lr: 0.001
     weight_decay: 0.00001
 
+# scheduler:
+  # name: null
+  # param: null
+
 scheduler:
-  name: null
-  param: null
-  # name: OneCycleLR
-  # param: 
-  # pct_start: 0.1
-  # div_factory: 0.001
-  # max_lr: 0.01
-  # epochs: 3
+  name: ExponentialLR
+  param:
+    gamma: 0.1
+    last_epoch: -1
+    verbose: false
+
+# scheduler:
+#   name: OneCycleLR
+#   param: 
+#     pct_start: 0.1
+#     div_factor: 0.001
+#     max_lr: 0.01
+#     epochs: 3
 
 loss_function:
-  name: SmoothBCEwLogits
-  param:
-    smoothing: 0.005
+  # name: SmoothBCEwLogits
+  # param:
+  #   smoothing: 0.005
+  name: BCEWithLogitsLoss
+  param: null
